refactor(aws-lambda): merge custom-context into types so they are excluded from coverage
