+++ src/adapter/aws-lambda/custom-context.ts
@@ -1,87 +1,20 @@
-interface ClientCert {
-  clientCertPem: string
-  subjectDN: string
-  issuerDN: string
-  serialNumber: string
-  validity: {
-    notBefore: string
-    notAfter: string
-  }
-}
+import type {
+  ALBRequestContext as ALBRequestContext_,
+  ApiGatewayRequestContextV2 as ApiGatewayRequestContextV2_,
+  ApiGatewayRequestContext as ApiGatewayRequestContext_,
+} from './types'
 
-interface Identity {
-  accessKey?: string
-  accountId?: string
-  caller?: string
-  cognitoAuthenticationProvider?: string
-  cognitoAuthenticationType?: string
-  cognitoIdentityId?: string
-  cognitoIdentityPoolId?: string
-  principalOrgId?: string
-  sourceIp: string
-  user?: string
-  userAgent: string
-  userArn?: string
-  clientCert?: ClientCert
-}
+/**
+ * @deprecated Use `ApiGatewayRequestContext` from `@src/adapter/aws-lambda/types` instead.
+ */
+export type ApiGatewayRequestContext = ApiGatewayRequestContext_
 
-export interface ApiGatewayRequestContext {
-  accountId: string
-  apiId: string
-  authorizer: {
-    claims?: unknown
-    scopes?: unknown
-  }
-  domainName: string
-  domainPrefix: string
-  extendedRequestId: string
-  httpMethod: string
-  identity: Identity
-  path: string
-  protocol: string
-  requestId: string
-  requestTime: string
-  requestTimeEpoch: number
-  resourceId?: string
-  resourcePath: string
-  stage: string
-}
+/**
+ * @deprecated Use `ApiGatewayRequestContextV2` from `hono/aws-lambda/types` instead.
+ */
+export type ApiGatewayRequestContextV2 = ApiGatewayRequestContextV2_
 
-interface Authorizer {
-  iam?: {
-    accessKey: string
-    accountId: string
-    callerId: string
-    cognitoIdentity: null
-    principalOrgId: null
-    userArn: string
-    userId: string
-  }
-}
-
-export interface ApiGatewayRequestContextV2 {
-  accountId: string
-  apiId: string
-  authentication: null
-  authorizer: Authorizer
-  domainName: string
-  domainPrefix: string
-  http: {
-    method: string
-    path: string
-    protocol: string
-    sourceIp: string
-    userAgent: string
-  }
-  requestId: string
-  routeKey: string
-  stage: string
-  time: string
-  timeEpoch: number
-}
-
-export interface ALBRequestContext {
-  elb: {
-    targetGroupArn: string
-  }
-}
+/**
+ * @deprecated Use `ApiGatewayRequestContext` from `@src/adapter/aws-lambda/types` instead.
+ */
+export type ALBRequestContext = ALBRequestContext_
