+++ src/adapter/aws-lambda/index.ts
@@ -9,5 +9,5 @@ export type {
   ApiGatewayRequestContext,
   ApiGatewayRequestContextV2,
   ALBRequestContext,
-} from './custom-context'
-export type { LambdaContext } from './types'
+  LambdaContext,
+} from './types'
